+++ src/AURORA.jl
@@ -12,8 +12,11 @@ include("ionosphere/ionosphere.jl")
 export Ionosphere
 include("ionosphere/iri/iri.jl")
 include("ionosphere/msis/msis.jl")
+include("ionosphere/msis/neutral_profile.jl")
 export find_msis_file, find_nrlmsis_file
+export NeutralProfile, DensityProfile, run_msis, read_msis_file, read_ccmc_msis
 export find_iri_file
+export ElectronProfile, run_iri, read_iri_file, read_ccmc_iri
 
 include("physics/cross_sections/e_N2_cross_sections.jl")
 include("physics/cross_sections/e_O2_cross_sections.jl")
@@ -37,7 +40,7 @@ include("physics/cascading_cache.jl")
 export clear_cascading_cache!
 
 include("physics/species.jl")
-export NeutralSpecies, MSISDensity, VectorDensity
+export NeutralSpecies
 export N2Species, O2Species, OSpecies
 
 include("model.jl")